#851 Move Serialization Interface into Abstractions (#858)
* Tweaks


* Add missing dependencies


* switch to system.memory


* Added missing using


* Reduce csproj file contents


---------
