+++ src/NATS.Client.Abstractions/NATS.Client.Abstractions.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <!-- NuGet Packaging -->
+    <PackageTags>pubsub;messaging;abstractions</PackageTags>
+    <Description>NATS client Abstractions for .NET</Description>
+  </PropertyGroup>
+
+  <!--  Dependencies for all -->
+  <ItemGroup>
+    <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.10.48">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="System.Memory" Version="4.5.5"/>
+  </ItemGroup>
+
+</Project>
